+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/backoffice-v2
 
+## 0.7.137
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.115
+  - @ballerine/common@0.9.92
+  - @ballerine/workflow-node-sdk@0.6.115
+
+## 0.7.136
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.136
+  - @ballerine/react-pdf-toolkit@1.2.107
+
 ## 0.7.135
 
 ### Patch Changes