+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,22 @@
 # kyb-app
 
+## 0.3.167
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.115
+  - @ballerine/common@0.9.92
+
+## 0.3.166
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/ui@0.7.136
+
 ## 0.3.165
 
 ### Patch Changes