+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-core
 
+## 0.6.115
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.92
+
 ## 0.6.114
 
 ### Patch Changes